cuda: captures the state of splitting C99 and C++ interfaces for contexts and arguments
